Criando usuário ADMIN padrão para o sistema

+++ curso-jsp/src/main/java/dao/DAOUsuarioRepository.java
@@ -54,7 +54,7 @@ public List<ModelLogin> consultaUsuarioList() throws Exception {
 		
 		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
 		
-		String sql = "select * from model_login ";
+		String sql = "select * from model_login where useradmin is false";
 		PreparedStatement statement = connection.prepareStatement(sql);
 		
 		ResultSet resultado = statement.executeQuery();
@@ -81,7 +81,7 @@ public List<ModelLogin> consultaUsuarioList(String nome) throws Exception {
 		
 		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
 		
-		String sql = "select * from model_login  where upper(nome) like upper(?) ";
+		String sql = "select * from model_login  where upper(nome) like upper(?) and useradmin is false";
 		PreparedStatement statement = connection.prepareStatement(sql);
 		statement.setString(1, "%" + nome + "%");
 		
@@ -109,7 +109,7 @@ public List<ModelLogin> consultaUsuarioList(String nome) throws Exception {
 
 		ModelLogin modelLogin = new ModelLogin();
 
-		String sql = "select * from model_login where upper (login) = upper ('" + login + "')";
+		String sql = "select * from model_login where upper (login) = upper ('" + login + "') and useradmin is false";
 
 		PreparedStatement statement = connection.prepareStatement(sql); // preparando SQL
 
@@ -131,7 +131,7 @@ public List<ModelLogin> consultaUsuarioList(String nome) throws Exception {
 
 		ModelLogin modelLogin = new ModelLogin();
 
-		String sql = "select * from model_login where id = ? ";
+		String sql = "select * from model_login where id = ? and useradmin is false";
 
 		PreparedStatement statement = connection.prepareStatement(sql); // preparando SQL
 
@@ -163,7 +163,7 @@ public List<ModelLogin> consultaUsuarioList(String nome) throws Exception {
 	}
 	
 	public void deletarUser(String idUser) throws Exception{
-		String sql = "delete from model_login where id = ?;";
+		String sql = "delete from model_login where id = ? and useradmin is false;";
 		
 		PreparedStatement prepareSql = connection.prepareStatement(sql);
